Rename the ES preset to ESX and create a proper preset file for the ES rim #1
Fixes the key mappings to represent the buttons on the Moza ES wheel. The ones that came with the game are mapped for the ESX Xbox wheel.

Just the original ES wheel configuration, but renamed to ESX to represent the actual bindings for that wheel. The name will show up in game as lng_moza_rim_esx.